feat: replace build.js with build.mts using zx for improved build process
