+++ package.json
@@ -5,7 +5,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "node build.js",
+    "build": "tsx build.mts",
     "dev": "vite build --watch --config vite.config.content.ts",
     "test": "jest",
     "test:watch": "jest --watch",
